add inital zsh configuration
this is a rough port of my existing bash configuration, though not quite
as neatly organized because I'm not sure yet the best way to do it yet.
I'm not actually sure if it makes sense to split all my functions into
individual files, especially given how short most of them are, but we'll
see.