vim: speed improvements when editing go files
Run gofmt on save rather than goimports, but run goimports on
<LocalLeader>f.  This reverses the current behavior and is designed to
use the faster of the two on save, and the (slightly) slower only when
requested.  This also addresses a problem where goimports would
occasionally get the import wrong, but it was being automatically run
every time the file was written.

Don't automatically run the syntastic checkers, instead only running
them on command.  Again, this is done to make writing files much faster.