vim: add local symlinks for some vim-go files
vim8's package loading code seems to put scripts in a different order
depending on whether ~/.vim/pack/{} is a symlink or not.  With it being
a symlink, the system go syntax file was getting used instead of the one
provided by vim-go.  Adding these symlinks directly in my ~/.vim
directory seems to fix this.  I suspect there is a better solution, but
this works for now.

I suspect this is the same bug that was fixed in
https://github.com/vim/vim/commit/2f9e575.