nvim: update keymaps for trouble
use 'x' for trouble a little more consistently... ctrl-x to open
telescope results in trouble for example. Similarly, use [x and ]x to
navigate between diagnostics in a buffer. Of course, diagnostics aren't
specific to trouble, but it's mostly what I use trouble for. Also switch
to [d and ]d to navigate between diffs. Using [c and ]c for "change"
just wasn't sticking in my head, but "d for diff" is easier to remember.