zsh: fix path ordering on macOS
There is a long history of macOS messing up ordering of elements in
$PATH due to the path_helper utility, which is pretty well summarized at
https://gist.github.com/Linerre/f11ad4a6a934dcf01ee8415c9457e7b2. For
zsh, path_helper is called by the system /etc/zprofile, which is loaded
after zshenv and before zshrc.

If the value of $PATH changes between loading zshenv and zshrc, we
source the local zshenv again so that any PATH elements can be nudged to
the front again.  Note that this is after calling `typset -U path` to
ensure entries aren't duplicated.

This is not the first time trying to sort this out, and very well may
not be the last, but it seems to be working pretty well for now.