nvim: update plugins
unset statuscolumn option, which is a new neovim feature that LazyVim
added support for in https://github.com/LazyVim/LazyVim/commit/364bcf3.
In particular, LazyVim caused the git line indicators to move to the
right of the line number, which I really don't like. I can probably find
a way to customize that without unsetting statuscolumn entirely, but
that's a problem for another day.

Also rename null-ls to the none-ls community fork.