nvim: split editorconfig for nvim lua files
For some reason, nvim doesn't seem to be loading the lua config from my
existing .editorconfig file. Only when I put it in my config/nvim
directory (or lower) does it actually load. This used to work, so seems
to be some kind of regression, but in the meantime just put the
lua-specific config options in this separate file.