hyprland: re-enable cursor warp
This seems to help reduce accidentally changing active windows by my
palm touching the trackpad, since the cursor is in the middle of the
active window.

I think the reason I disabled it is that my palm would result in
highlighting text in the active window, which may end up being a problem
again.