nvim: fix Super mod key on linux
Previously, I was using <M-*> for a handful of key maps, which maps to
the Alt key on Linux, and Option on macOS. What I really want is to use
Super on Linux, since that is in the same physical placement as mac's
Option key, and because I use Alt on linux mostly for window management
and system-level shortcuts (similar to Command on mac).

I didn't seem to have issue with this until recently. I'm honestly not
sure how it even worked until now; it probably shouldn't have. I'm
guessing something changed in a recent built of neovim to change this.

In any event, this adds a "util.mod" function that returns a "<M-*>"
keybinding string on mac, and "<D-*>" on all other platforms. That way,
I don't have to try to do weird key mapping changes in my terminal apps
or elsewhere, which didn't seem to work very well anyway.