sway: don't start Firefox on launch
I'm not completely sure it's due to starting Firefox on launch, but I've
had a number of instances lately where Firefox loses all of my open tabs
on restart.

I *think* what is happening is that firefox doesn't fully start until I
switch to workspace 2. Maybe it starts but doesn't reopen windows and
tabs? So if I start sway, and then exit without switching to workspace
2, then firefox shuts down thinking that there aren't any tabs or
windows to save. So on the next launch, the tabs are lost.