waybar: disable fixed-center instead of limiting sway/window length
This causes window titles to be off center in waybar, but is a much
better use of space and actually looks fine. This also seems to give
preference to modules-right, truncating the sway/window text before
truncating module icons in modules-right, which was really what I was
trying to do with max-length anyway.