sway: swap left super and alt keys, use super as modifier
I like to use the key left of the space bar (normally Alt) for key binds
in my window manager, and the next key over (normally Super) for the
occasional application-specific key binds (mostly some stuff in neovim).
Most of my neovim keybinds use Ctrl, and only a small handful use the
Super key. However, I use sway key bindings ALL the time to switch
windows and workspaces, so I like having it use a more centrally located
modifier key.

I was having the strangest time in neovim where sometimes the Super key
would send an Alt modifier, and sometimes it would send a normal Super
modifier (see 395a1ee and 33062f1). I could never track down why it
switched until today; it depends on whether I'm in tmux or not. Outside
of tmux, it sends a normal Super modifier, since both my terminal and
neovim support the Super key just fine. But tmux does not, so it sends
an Alt modifier instead. I'm in a tmux session 90+% of the time I'm
using neovim, so it was only the occasional time I wasn't in tmux that I
would get different behavior.

To fix this, I'm changing my keyboard layout to swap the left Alt and
Super keys, then updating all of my sway bindings to use Super instead
of Alt. That way it's the same physical key, but just sends a different
modifier. And then I can use Alt for my neovim bindings, and it works
consistently both in and out of tmux. This also happens to map more
closely to the layout on a Mac keyboard, which is both probably why I
prefer it (20 years of muscle memory), and it just seems to make sense
to me.

I've also gone so far as to switch the physical key caps on my laptop
and my external keyboard to match the key swap.