linux: begin to add support for light theme
I'm finding my default theme (in kitty in particular) to be difficult to
see in direct sunlight.  This change begins to add a light mode for
kitty, and uses the Framework logo key to cycle through color schemes.

Kitty colors for light mode are from https://github.com/sonph/onehalf.

I'm not completely happy with things, and need to add support for
enabling light mode in onedark.nvim, but this is a start.