headless-output-start: move current container to new output
When creating a new headless output, it's (currently) always because I'm
going to share one or more application windows. Typically, that is the
active window at the time that I decide I want to share. So go ahead and
move the current container to the newly created headless output. On the
rare occasion that isn't the window I wanted, I can always move it back,
but I'm pretty sure this is going to make screen sharing a tad easier.