tailscale: add wrapper script for `tailscale systray`
We've recently added the `systray` subcommand to the tailscale CLI.
This means I'd either need to use `pgrep -f` or `pkill -f` to either
find or kill the process, since ps or pkill only search the process name
without arguments.  For whatever reason, I can't get `sway exec`
commands to handle pgrep or pkill properly. So instead of fighting with
that, it's a lot easier to just add a simple wrapper script.