sway: add new '-' workspace for screensharing, update keybinds
Instead of using workspace 9 for screensharing, which I was previously
doing and is beginning to regularly conflict with using that as a normal
workspace, create a new dedicated '-' workspace just for screensharing.

Add keybinds to switch and move windows to this new workspace just like
others, so `Alt+Minux` switches to it and `Alt+Shift+Minus` moves a
window there. This requires shifting a few other keybinds:

 - showing the scratchpad is now either `Alt+Grave` or `Alt+Equal`. I
   don't know which I'll like better, so I'm adding both for now.
 - showing the notification center is now `Alt+Backspace`. I'm not sure
   if I'll like this, but trying it for now.

Enabling and disabling the headless output for screensharing now also
uses the minus key: `Alt+Super+Minus` to enable and
`Alt+Super+Shift+Minus` to disable.