z: use __z as the Z command
I was exploring using the bare z command as a wrapper for zellij,
similar to my ts wrapper for tmux. I'm probably not going to pursue that
right now, but I was never actually using the z command anyway. Instead,
I use the j command which wraps z with an fzf picker. So this just frees up z for future use.