nvim.mini: add snacks and enable snacks.picker
I'll try to use mini.nvim for most things on this config, but I really
like the preview window of snacks.picker and am not willing to give that
up.  I think there will probably be a few other things in snacks.nvim
that I'll end up pulling in anyway.